Add navigation links to the top of recipe pages
Styles are now added to make the page a bit more easier to navigate.
Also, styles that are no longer useful have been removed. The layout now
follows the HTML5 standard a little bit more now. This update came as a
result of learning and understand flex box to some degree.

The navigation links added to the top of the recipe pages lead to the
homepage and an about page which is yet to be written. In the future,
the number of links may increase to meet needs and are subject to change
in the long run as needed.